+++ History.md
@@ -2,6 +2,7 @@ Unreleased
 ==================
 
 * added; `applies_to_non_plan_charges` attribute `Coupon`
+* added; `Coupons` to `Subscription`
 
 1.2.2 (stable) / 2015-07-06
 ==================